Updated example to working
I had to dig in to the code to figure out why I was getting Not connected message. Turns out autodetect does not work very well therefore authmethod => 1 is requiredf or a more reliable means of connection. 

The example was also missing $tc->connect() call which is require in order to do $tc->authenticate().