Fix error in the readme; `server` should be `hostname`.
According to
dunglas/php-torcontrol@92e9f890ae66287e8e8b32d99bf9177cf59d05d8, when
invoking `connect()` the method checks for a `hostname`, not a `server`
key in the argument array. The readme should match this usage so as to
avoid likely confusion (and an unpleasant `IOError` exception). See line
number 174 at the above-mentioned commit for the exact details.