Fixes #7: Support multi-line replies from Tor.
This commit fixes an issue where a `ProtocolError` exception was
prematurely thrown before the end of a multi-line reply from Tor. The
issue was caused by ignoring the multi-line nature of certain replies
from Tor and always assuming single-line replies. This patch addresses
the issue by retaining state information about whether the current line
is a multi-line or single line reply.

This patch also fixes a typo in an error message (`Unknow`-> `Unknown`).

This patch includes unit tests that demonstrate the problem and the fix.